+++ bioblend/_tests/TestCloudmanLaunch.py
@@ -1,8 +1,6 @@
 """
 Tests the functionality of the Blend CloudMan API. These tests require working
 credentials to supported cloud infrastructure.
-
-Use ``nose`` to run these unit tests.
 """
 from bioblend.cloudman import CloudManConfig, CloudManInstance
 from . import CloudmanTestBase, test_util