+++ bioblend/_tests/TestGalaxyDatasets.py
@@ -1,6 +1,3 @@
-"""
-Use ``nose`` to run these unit tests.
-"""
 import shutil
 import tempfile