+++ bioblend/_tests/TestGalaxyQuotas.py
@@ -1,6 +1,3 @@
-"""
-Use ``nose`` to run these unit tests.
-"""
 import uuid
 
 from . import GalaxyTestBase