+++ bioblend/_tests/TestGalaxyWorkflows.py
@@ -1,6 +1,3 @@
-"""
-Use ``nose`` to run these unit tests.
-"""
 import json
 import os
 import shutil